introduce threading
threading :S this is a scary one, but i think it will be ok. The
'listener' thread will act as a listener and respond on verious queries
from one of us. the 'timer' thread will act on its own, and will send
massages based on the current time (eg. a good morning message including
the weather forcast).

this will be interesting.